# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/custom_field/custom_field/{id}": {
      "get": {
        "operationId": "CustomFieldController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsCustomFieldDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field"
        ]
      },
      "put": {
        "operationId": "CustomFieldController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsCustomFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for put ToolsCustomFieldDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field"
        ]
      },
      "delete": {
        "operationId": "CustomFieldController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for delete ToolsCustomFieldDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field"
        ]
      }
    },
    "/auth/custom_field_value/custom_field_value": {
      "get": {
        "operationId": "CustomFieldValueController_findAll",
        "parameters": [
          {
            "name": "pageIndex",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "JSON.stringify({})",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortKey",
            "required": false,
            "in": "query",
            "schema": {
              "example": "created_at",
              "type": "string"
            }
          },
          {
            "name": "sortValue",
            "required": false,
            "in": "query",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search_field",
            "required": false,
            "in": "query",
            "description": "Array of searchable fields",
            "schema": {
              "example": [
                "nama",
                "alamat"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "search_keyword",
            "required": false,
            "in": "query",
            "schema": {
              "example": "makmur",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for list ToolsCustomFieldValueDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ToolsCustomFieldValueDto"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 123
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field_value"
        ]
      },
      "post": {
        "operationId": "CustomFieldValueController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsCustomFieldValueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create ToolsCustomFieldValueDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldValueDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field_value"
        ]
      }
    },
    "/auth/custom_field_value/custom_field_value/{id}": {
      "get": {
        "operationId": "CustomFieldValueController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsCustomFieldValueDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldValueDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field_value"
        ]
      },
      "put": {
        "operationId": "CustomFieldValueController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsCustomFieldValueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for put ToolsCustomFieldValueDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldValueDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field_value"
        ]
      },
      "delete": {
        "operationId": "CustomFieldValueController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for delete ToolsCustomFieldValueDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldValueDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field_value"
        ]
      }
    },
    "/auth/custom_field_value/custom_field_value_report/list": {
      "post": {
        "operationId": "CustomFieldValueReportController_findAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseQueryDtoSmart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for list ToolsCustomFieldValueReportDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ToolsCustomFieldValueReportDto"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 123
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field_value_report"
        ]
      }
    },
    "/auth/custom_field_value/custom_field_value_report/{id}": {
      "get": {
        "operationId": "CustomFieldValueReportController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsCustomFieldValueReportDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsCustomFieldValueReportDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "custom_field_value_report"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP API",
    "description": "Dokumentasi API ERP multi-tenant",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BaseQueryDtoSmart": {
        "type": "object",
        "properties": {
          "filter": {
            "type": "object",
            "description": "Where filter (supports >, >=, <, <=, like, notLike,  in, notIn, isNull, isNotNull)",
            "example": {
              "id": {
                "in": [
                  "id1",
                  "id2"
                ]
              },
              "status": {
                "eq": "aktif"
              },
              "created_at": {
                "between": [
                  "2024-01-01",
                  "2024-12-31"
                ]
              }
            },
            "additionalProperties": true
          },
          "include": {
            "type": "array",
            "description": "Nama tabel beserta tipenya single atau array",
            "example": [
              {
                "name": "contract",
                "type": "single",
                "select": [
                  "idContract",
                  "idClient",
                  "nomorKontrak",
                  "statusKontrak"
                ]
              },
              {
                "name": "client",
                "parent": "contract",
                "type": "single"
              },
              {
                "name": "work_schedule",
                "type": "array"
              }
            ],
            "additionalProperties": true
          },
          "joinWhere": {
            "type": "array",
            "description": "Filter + jenis JOIN (inner/left/right)",
            "example": [
              {
                "contract": {
                  "status_kontrak": "aktif"
                },
                "type": "inner"
              },
              {
                "client_site": {
                  "kota": "Jakarta"
                },
                "type": "left"
              }
            ]
          },
          "search_field": {
            "description": "Fields to search with search_keyword",
            "example": [
              "contract.nomor_kontrak",
              "site.kota"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "search_keyword": {
            "type": "string",
            "description": "Search keyword",
            "example": "PT"
          },
          "sortKey": {
            "type": "string",
            "description": "Sort by field",
            "example": "contract.nomor_kontrak"
          },
          "sortValue": {
            "type": "string",
            "description": "Sort direction (asc or desc)",
            "example": "asc"
          },
          "pageIndex": {
            "type": "number",
            "description": "Pagination page index",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "Pagination limit per page",
            "example": 20
          }
        }
      },
...
      "ToolsCustomFieldDto": {
        "type": "object",
        "properties": {
          "idCustomField": {
            "type": "string"
          },
          "idCustomFieldGroup": {
            "type": "string"
          },
          "kodeField": {
            "type": "string"
          },
          "labelField": {
            "type": "string"
          },
          "prefixField": {
            "type": "string"
          },
          "tipeField": {
            "type": "string"
          },
          "opsiSelect": {
            "type": "string"
          },
          "isFilterable": {
            "type": "boolean"
          },
          "isRequired": {
            "type": "boolean"
          },
          "validasi": {
            "type": "string"
          },
          "defaultValue": {
            "type": "string"
          },
          "catatan": {
            "type": "string"
          },
          "urutan": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "idCustomField",
          "kodeField",
          "labelField",
          "tipeField"
        ]
      },
      "ToolsCustomFieldValueDto": {
        "type": "object",
        "properties": {
          "idCustomFieldValue": {
            "type": "string"
          },
          "idCustomField": {
            "type": "string"
          },
          "idCustomFieldGroup": {
            "type": "string"
          },
          "modul": {
            "type": "string"
          },
          "tableName": {
            "type": "string"
          },
          "modulId": {
            "type": "string"
          },
          "nilai": {
            "type": "string"
          },
          "nilaiJson": {
            "type": "string"
          },
          "kodeField": {
            "type": "string"
          },
          "labelField": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "idCustomFieldValue",
          "idCustomField",
          "modulId"
        ]
      },
      "ToolsCustomFieldValueReportDto": {
        "type": "object",
        "properties": {
          "idCustomFieldValue": {
            "type": "string"
          },
          "idCustomField": {
            "type": "string"
          },
          "idCustomFieldGroup": {
            "type": "string"
          },
          "modul": {
            "type": "string"
          },
          "tableName": {
            "type": "string"
          },
          "modulId": {
            "type": "string"
          },
          "nilai": {
            "type": "string"
          },
          "nilaiJson": {
            "type": "string"
          },
          "kodeField": {
            "type": "string"
          },
          "labelField": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "idCustomFieldValue",
          "idCustomField",
          "modulId"
        ]
      },
...
    }
  }
}